add non-nrt test for C#
