+++ PresqueYaml.CSharp.Tests/PresqueYaml.CSharp.Tests.csproj
@@ -7,6 +7,7 @@
 
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
+    <RootNamespace>PresqueYaml.CSharp.Tests</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
